refactor: Move some Android configuration and utilities around
