+++ src/configuration.js
@@ -1,0 +1,6 @@
+const config = {
+  quantumSpreadsheetId: '1UMsy_sZkdgtElr2buwRtABuyA3GY6wNK_pfF01c890A',
+  androidSpreadsheetId: '1vE0b3tawWY9vVNq9Ds6CiA9XZ6LStkcXZl3F8dwXid8',
+};
+
+export default config;
